pyproject.toml
[project]
name = "ibn-platform"
version = "0.1.0"
description = "Intent-Based Networking Platform - Portfolio Project"
requires-python = ">=3.11"
dependencies = [
    "z3-solver>=4.12",
    "networkx>=3.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "jinja2>=3.0",
    "netmiko>=4.6.0",
]

[project.scripts]
ibn = "ibn.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ibn"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.1",
    "mypy>=1.8",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
